fix(frontend): the unsaved guard sits above the screens
The navigation guard moves out of the settings screen and into App, above the
routed screen. Inside a screen it could only see the moves it stayed mounted
for: a settings draft was safe from one entry to the next and still discarded
without a word the moment the reader clicked anything else, because
SettingsScreen unmounted before the guard could ask. The address it holds is the
route hash, so `routeHash`/`parseHash` do the serialising and there is no second
spelling of an address. What waits is the subtree, never the URL — the browser's
own Back and Forward are still the browser's.

A conformance gate holds it there: exactly one file may render UnsavedGuard, and
a second one is by definition a guard that unmounts with its own screen. Two
Playwright cases carry the behaviour, because where the guard SITS is only
visible across a whole navigation — both fail when the address it holds is
frozen.

The toast suite drives its clicks through `userEvent` on a controlled clock, as
frontend/CLAUDE.md asks for a timer-driven component. That combination
deadlocked: testing-library drains its microtask queue by awaiting a zero-delay
timeout and only advances the clock alongside it when a `jest` global exists, so
every awaited interaction hung until the test timed out. The one method it
reaches for is provided, per file rather than globally — a suite whose clock
advances on its own must not have timers fired underneath it.

settings.test.tsx is split by subject into seven suites and a testkit, all well
under the 1000-line ceiling, with the same 76 tests and the same names.
